refactor: compute in double, narrow to float only at output
Every accumulator in the library builds up over hundreds of steps -- arc
length swept cell by cell, the profiler's time and arc-length state, the
RAMSETE pose integrated one dt at a time, the Newton/quadrature
intermediates -- and float rounding was visible in all of them. Move the
shared value types (types.hpp) and every internal variable, member,
parameter and tolerance to double.

float now appears in exactly one place: printer.cpp, the output boundary,
where the double state is narrowed on the way out because a Desmos plot or a
robot's velocity controller only consumes float precision.

Newton and cusp tolerances are retightened to match double (1e-6f -> 1e-9,
1e-12f -> 1e-15, etc.); at float precision those thresholds were near the
noise floor. No test tolerance needed changing -- the existing suite is the
proof that 2D output is unchanged to within its stated bounds; double only
tightens the numbers. The two test edits are mechanical: projectOntoCurve's
residual out-param and a std::max accumulator follow the new return types.

+++ include/bezier.hpp
@@ -19,23 +19,23 @@ public:
 // routine or to a profile.
 void requireCubicSegment(const std::vector<Point>& controlPoints);
 
-Point bezierDerivative(const std::vector<Point>& controlPoints, float t);
-Point bezierSecondDerivative(const std::vector<Point>& controlPoints, float t);
-float speed(const std::vector<Point>& controlPoints, float t);
-float arcLength(const std::vector<Point>& controlPoints, float a, float b);
-float sFunction(const std::vector<Point>& controlPoints, float t);
+Point bezierDerivative(const std::vector<Point>& controlPoints, double t);
+Point bezierSecondDerivative(const std::vector<Point>& controlPoints, double t);
+double speed(const std::vector<Point>& controlPoints, double t);
+double arcLength(const std::vector<Point>& controlPoints, double a, double b);
+double sFunction(const std::vector<Point>& controlPoints, double t);
 // tGuess seeds the Newton iteration; pass the previous parameter when stepping
 // along a path so the solver starts within one timestep of the answer.
-float findTForS(const std::vector<Point>& controlPoints, float sCurrent, float deltaS,
-                float tGuess = 0.5f);
+double findTForS(const std::vector<Point>& controlPoints, double sCurrent, double deltaS,
+                 double tGuess = 0.5);
 
-Pose findXandY(const std::vector<Point>& controlPoints, float t);
+Pose findXandY(const std::vector<Point>& controlPoints, double t);
 
 // Returns the t minimising ||r(t) - (x, y)||. If residual is non-null it
 // receives the distance from the curve to (x, y), which is how a caller tells
 // an on-path point from one that merely projects somewhere.
-float projectOntoCurve(const std::vector<Point>& controlPoints, float x, float y,
-                       float* residual = nullptr);
+double projectOntoCurve(const std::vector<Point>& controlPoints, double x, double y,
+                        double* residual = nullptr);
 
 // Thrown for a keyframe that cannot be placed on the path: either too far from
 // the curve to be meaningful, or out of order along it.
@@ -50,5 +50,5 @@ std::vector<KeyframeVelocities> convertToTFrame(
     const std::vector<KeyframeVelocitiesXandY>& keyFrameVelocitiesXY
 );
 
-float signedCurvature(const std::vector<Point>& controlPoints, float t);
-float unsignedCurvature(const std::vector<Point>& controlPoints, float t);
+double signedCurvature(const std::vector<Point>& controlPoints, double t);
+double unsignedCurvature(const std::vector<Point>& controlPoints, double t);
